Add more tests to core. (#315)
* Add nock.test.ts.

* Clean tests.

* Simplify a bit.

* Remove duplicate tests.
