+++ packages/unmock-core/src/__tests__/corePackage.test.ts
@@ -1,21 +1,5 @@
 import { UnmockPackage } from "..";
-import Backend from "../backend/";
-
-class TestBackend extends Backend {
-  public constructor() {
-    super({ InterceptorCls: jest.fn() });
-  }
-  public initialize(_: any): never {
-    throw Error("Not implemented");
-  }
-  public reset() {
-    return;
-  }
-  public get services() {
-    return {};
-  }
-  public loadServices() {}
-}
+import { TestBackend } from "./utils";
 
 const backend = new TestBackend();
